Rename and use the 1.55 ABI for 1.54
