+++ crates/proc_macro_srv/src/abis/abi_1_54/proc_macro/bridge/scoped_cell.rs
