+++ crates/proc_macro_srv/src/abis/mod.rs
@@ -25,13 +25,13 @@
 
 // pub(crate) so tests can use the TokenStream, more notes in test/utils.rs
 pub(crate) mod abi_1_47;
-mod abi_1_55;
+mod abi_1_54;
 mod abi_1_56;
 mod abi_1_58;
 
 use super::dylib::LoadProcMacroDylibError;
 pub(crate) use abi_1_47::Abi as Abi_1_47;
-pub(crate) use abi_1_55::Abi as Abi_1_55;
+pub(crate) use abi_1_54::Abi as Abi_1_54;
 pub(crate) use abi_1_56::Abi as Abi_1_56;
 pub(crate) use abi_1_58::Abi as Abi_1_58;
 use libloading::Library;
@@ -49,7 +49,7 @@ impl PanicMessage {
 
 pub(crate) enum Abi {
     Abi1_47(Abi_1_47),
-    Abi1_55(Abi_1_55),
+    Abi1_54(Abi_1_54),
     Abi1_56(Abi_1_56),
     Abi1_58(Abi_1_58),
 }
@@ -72,13 +72,13 @@ impl Abi {
         // FIXME: this should use exclusive ranges when they're stable
         // https://github.com/rust-lang/rust/issues/37854
         match (info.version.0, info.version.1) {
-            (1, 47..=54) => {
+            (1, 47..=53) => {
                 let inner = unsafe { Abi_1_47::from_lib(lib, symbol_name) }?;
                 Ok(Abi::Abi1_47(inner))
             }
-            (1, 55..=55) => {
-                let inner = unsafe { Abi_1_55::from_lib(lib, symbol_name) }?;
-                Ok(Abi::Abi1_55(inner))
+            (1, 54..=55) => {
+                let inner = unsafe { Abi_1_54::from_lib(lib, symbol_name) }?;
+                Ok(Abi::Abi1_54(inner))
             }
             (1, 56..=57) => {
                 let inner = unsafe { Abi_1_56::from_lib(lib, symbol_name) }?;
@@ -99,8 +99,8 @@ impl Abi {
         attributes: Option<&tt::Subtree>,
     ) -> Result<tt::Subtree, PanicMessage> {
         match self {
-            Self::Abi1_55(abi) => abi.expand(macro_name, macro_body, attributes),
             Self::Abi1_47(abi) => abi.expand(macro_name, macro_body, attributes),
+            Self::Abi1_54(abi) => abi.expand(macro_name, macro_body, attributes),
             Self::Abi1_56(abi) => abi.expand(macro_name, macro_body, attributes),
             Self::Abi1_58(abi) => abi.expand(macro_name, macro_body, attributes),
         }
@@ -109,7 +109,7 @@ impl Abi {
     pub fn list_macros(&self) -> Vec<(String, ProcMacroKind)> {
         match self {
             Self::Abi1_47(abi) => abi.list_macros(),
-            Self::Abi1_55(abi) => abi.list_macros(),
+            Self::Abi1_54(abi) => abi.list_macros(),
             Self::Abi1_56(abi) => abi.list_macros(),
             Self::Abi1_58(abi) => abi.list_macros(),
         }
